+++ src/main/java/de/samply/frontend/dto/ProjectDocument.java
@@ -14,6 +14,7 @@ public record ProjectDocument(
         String bridgehead,
         String humanReadableBridgehead,
         String creatorEmail,
+        String creatorName,
         String label,
         DocumentType type
 ) {